+++ docs/changes.rst
@@ -6,6 +6,9 @@ Version 0.4.0
 
 To be released.
 
+- Added :class:`~wikidata.multilingual.MonolingualText` type which is a true
+  subtype of :class:`str`.
+
 
 Version 0.3.0
 '''''''''''''